+++ lib/tensor_stream/nn/nn_ops.rb
@@ -1,7 +1,10 @@
+require 'tensor_stream/nn/embedding_lookup'
 module TensorStream
   # High level machine learning functions
   class NN
     extend TensorStream::OpHelper
+    extend TensorStream::EmbeddingLookup
+    extend TensorStream::Maths::MathFunctions
 
     class << self
       def softmax(logits, axis: nil, name: nil)