+++ lib/tensor_stream/ops/sum.rb
@@ -7,14 +7,16 @@ TensorStream::OpMaker.define_operation :sum do |op|
   op.what_it_does "If axis has no entries, all dimensions are reduced, and a tensor with a single element is returned."
 
   op.parameter :input_a, "tensor X"
-  op.parameter :axis, "tensor X", :nil, validate: 'INTEGER_TYPES'
+  op.parameter :axis_p, "tensor X", :nil, validate: 'INTEGER_TYPES'
 
+  op.option :axis, "axis", :nil, exclude: true
   op.option :name, "Optional name", :nil
   op.option :keepdims, "If true, retains reduced dimensions with length 1.", :false
 
   op.add_custom "input_a = TensorStream.convert_to_tensor(input_a)"
   op.add_custom "return input_a if input_a.shape.scalar?"
-  op.add_custom "axis = cast_axis(input_a, axis)"
+  op.add_custom "axis_p = axis_p || axis"
+  op.add_custom "axis_p = cast_axis(input_a, axis_p)"
 
   op.define_gradient do |grad, node, params|
     x, y = params